+++ source/sites/publicodes/Simulateur.js
@@ -13,7 +13,9 @@ import { Helmet } from 'react-helmet'
 import { useDispatch, useSelector } from 'react-redux'
 import { Redirect } from 'react-router'
 import CarbonImpact from './CarbonImpact'
-import Chart, { extractCategories } from './chart/index.js'
+import Chart from './chart/index.js'
+import { extractCategories } from 'Components/publicodesUtils'
+
 import { objectifsSelector } from 'Selectors/simulationSelectors'
 import { useEngine } from 'Components/utils/EngineContext'
 import emoji from 'react-easy-emoji'
